Added KmsDefaultDecrypter
