+++ nimbus-jose-jwt_aws-kms-extension/src/main/java/com/nimbusds/jose/aws/kms/crypto/KmsSymmetricDecrypter.java
@@ -17,35 +17,17 @@
 package com.nimbusds.jose.aws.kms.crypto;
 
 import com.amazonaws.services.kms.AWSKMS;
-import com.amazonaws.services.kms.model.DecryptRequest;
-import com.amazonaws.services.kms.model.DecryptResult;
-import com.amazonaws.services.kms.model.DependencyTimeoutException;
-import com.amazonaws.services.kms.model.DisabledException;
-import com.amazonaws.services.kms.model.InvalidGrantTokenException;
-import com.amazonaws.services.kms.model.InvalidKeyUsageException;
-import com.amazonaws.services.kms.model.KMSInternalException;
-import com.amazonaws.services.kms.model.KMSInvalidStateException;
-import com.amazonaws.services.kms.model.KeyUnavailableException;
-import com.amazonaws.services.kms.model.NotFoundException;
 import com.nimbusds.jose.CriticalHeaderParamsAware;
-import com.nimbusds.jose.EncryptionMethod;
 import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWEAlgorithm;
 import com.nimbusds.jose.JWEDecrypter;
 import com.nimbusds.jose.JWEHeader;
-import com.nimbusds.jose.RemoteKeySourceException;
 import com.nimbusds.jose.aws.kms.crypto.impl.KmsSymmetricCryptoProvider;
-import com.nimbusds.jose.aws.kms.exceptions.TemporaryJOSEException;
-import com.nimbusds.jose.crypto.impl.AlgorithmSupportMessage;
-import com.nimbusds.jose.crypto.impl.ContentCryptoProvider;
+import com.nimbusds.jose.aws.kms.crypto.utils.JWEDecrypterUtil;
 import com.nimbusds.jose.crypto.impl.CriticalHeaderParamsDeferral;
 import com.nimbusds.jose.util.Base64URL;
-import java.nio.ByteBuffer;
 import java.util.Map;
 import java.util.Set;
 import javax.annotation.concurrent.ThreadSafe;
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
 import lombok.NonNull;
 
 /**
@@ -106,26 +88,7 @@ public class KmsSymmetricDecrypter extends KmsSymmetricCryptoProvider implements
         validateJWEHeader(header);
         critPolicy.ensureHeaderPasses(header);
 
-        final DecryptResult cekDecryptResult =
-                decryptCek(getKeyId(), getEncryptionContext(), encryptedKey);
-        final SecretKey cek =
-                new SecretKeySpec(cekDecryptResult.getPlaintext().array(), header.getAlgorithm().toString());
-
-        return ContentCryptoProvider.decrypt(header, encryptedKey, iv, cipherText, authTag, cek, getJCAContext());
-    }
-
-    private DecryptResult decryptCek(String keyId, Map<String, String> encryptionContext, Base64URL encryptedKey)
-            throws JOSEException {
-        try {
-            return getKms().decrypt(new DecryptRequest()
-                    .withEncryptionContext(encryptionContext)
-                    .withKeyId(keyId)
-                    .withCiphertextBlob(ByteBuffer.wrap(encryptedKey.decode())));
-        } catch (NotFoundException | DisabledException | InvalidKeyUsageException | KeyUnavailableException
-                | KMSInvalidStateException e) {
-            throw new RemoteKeySourceException("An exception was thrown from KMS due to invalid key.", e);
-        } catch (DependencyTimeoutException | InvalidGrantTokenException | KMSInternalException e) {
-            throw new TemporaryJOSEException("A temporary error was thrown from KMS.", e);
-        }
+        return JWEDecrypterUtil.decrypt(getKms(), getKeyId(), getEncryptionContext(), header, encryptedKey, iv,
+                cipherText, authTag, getJCAContext());
     }
 }
