+++ nimbus-jose-jwt_aws-kms-extension/src/test/java/com/nimbusds/jose/aws/kms/crypto/KmsDefaultEncrypterTest.java
@@ -49,19 +49,15 @@ import com.nimbusds.jose.JWECryptoParts;
 import com.nimbusds.jose.JWEHeader;
 import com.nimbusds.jose.RemoteKeySourceException;
 import com.nimbusds.jose.aws.kms.crypto.impl.KmsDefaultEncryptionCryptoProvider;
-import com.nimbusds.jose.aws.kms.crypto.impl.KmsSymmetricCryptoProvider;
 import com.nimbusds.jose.aws.kms.crypto.testUtils.EasyRandomTestUtils;
 import com.nimbusds.jose.aws.kms.exceptions.TemporaryJOSEException;
 import com.nimbusds.jose.crypto.impl.ContentCryptoProvider;
-import com.nimbusds.jose.jca.JCAAware;
-import com.nimbusds.jose.jca.JCAContext;
 import com.nimbusds.jose.jca.JWEJCAContext;
 import com.nimbusds.jose.util.Base64URL;
 import java.nio.ByteBuffer;
 import java.security.SecureRandom;
 import java.util.Map;
 import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
 import lombok.SneakyThrows;
 import lombok.var;
 import org.jeasy.random.EasyRandom;
