+++ nimbus-jose-jwt_aws-kms-extension/src/test/java/com/nimbusds/jose/aws/kms/crypto/KmsSymmetricDecrypterTest.java
@@ -19,33 +19,21 @@ package com.nimbusds.jose.aws.kms.crypto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.amazonaws.services.kms.AWSKMS;
-import com.amazonaws.services.kms.model.AWSKMSException;
 import com.amazonaws.services.kms.model.DecryptRequest;
 import com.amazonaws.services.kms.model.DecryptResult;
-import com.amazonaws.services.kms.model.DependencyTimeoutException;
-import com.amazonaws.services.kms.model.DisabledException;
 import com.amazonaws.services.kms.model.EncryptionAlgorithmSpec;
-import com.amazonaws.services.kms.model.InvalidGrantTokenException;
-import com.amazonaws.services.kms.model.InvalidKeyUsageException;
-import com.amazonaws.services.kms.model.KMSInternalException;
-import com.amazonaws.services.kms.model.KMSInvalidStateException;
-import com.amazonaws.services.kms.model.KeyUnavailableException;
-import com.amazonaws.services.kms.model.NotFoundException;
 import com.google.common.collect.ImmutableSet;
 import com.nimbusds.jose.EncryptionMethod;
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWEAlgorithm;
 import com.nimbusds.jose.JWEHeader;
-import com.nimbusds.jose.RemoteKeySourceException;
 import com.nimbusds.jose.aws.kms.crypto.testUtils.EasyRandomTestUtils;
-import com.nimbusds.jose.aws.kms.exceptions.TemporaryJOSEException;
 import com.nimbusds.jose.crypto.impl.ContentCryptoProvider;
 import com.nimbusds.jose.crypto.impl.CriticalHeaderParamsDeferral;
 import com.nimbusds.jose.util.Base64URL;
@@ -54,7 +42,6 @@ import java.util.Map;
 import java.util.Set;
 import javax.crypto.spec.SecretKeySpec;
 import lombok.SneakyThrows;
-import lombok.var;
 import org.jeasy.random.EasyRandom;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -62,8 +49,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.junit.platform.commons.support.ReflectionSupport;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
@@ -80,7 +65,6 @@ class KmsSymmetricDecrypterTest {
     private String testKeyId;
     private Map<String, String> testEncryptionContext;
     private Set<String> testDeferredCriticalHeaders;
-
     private KmsSymmetricDecrypter kmsSymmetricDecrypter;
 
     @BeforeEach
@@ -192,72 +176,8 @@ class KmsSymmetricDecrypterTest {
             }
 
             @Nested
-            @DisplayName("with invalid key exception from KMS,")
-            class WithInvalidKMSKeyException {
-
-                AWSKMSException parameterizedBeforeEach(final Class<AWSKMSException> invalidKeyExceptionClass) {
-                    final var invalidKeyException = mock(invalidKeyExceptionClass);
-                    when(mockAwsKms
-                            .decrypt(new DecryptRequest()
-                                    .withEncryptionContext(testEncryptionContext)
-                                    .withKeyId(testKeyId)
-                                    .withCiphertextBlob(ByteBuffer.wrap(testEncryptedKey.decode()))))
-                            .thenThrow(invalidKeyException);
-
-                    return invalidKeyException;
-                }
-
-                @ParameterizedTest
-                @DisplayName("should throw RemoteKeySourceException.")
-                @ValueSource(classes = {
-                        NotFoundException.class, DisabledException.class, InvalidKeyUsageException.class,
-                        KeyUnavailableException.class, KMSInvalidStateException.class})
-                void shouldThrowRemoteKeySourceException(final Class<AWSKMSException> invalidKeyExceptionClass) {
-                    final var invalidKeyException = parameterizedBeforeEach(invalidKeyExceptionClass);
-                    assertThatThrownBy(
-                            () -> kmsSymmetricDecrypter.decrypt(testJweHeader, testEncryptedKey, testIv, testCipherText,
-                                    testAuthTag))
-                            .isInstanceOf(RemoteKeySourceException.class)
-                            .hasMessage("An exception was thrown from KMS due to invalid key.")
-                            .hasCause(invalidKeyException);
-                }
-            }
-
-            @Nested
-            @DisplayName("with a temporary exception from KMS,")
-            class WithTemporaryKMSException {
-
-                AWSKMSException parameterizedBeforeEach(final Class<AWSKMSException> temporaryKMSExceptionClass) {
-                    final var temporaryKMSException = mock(temporaryKMSExceptionClass);
-                    when(mockAwsKms
-                            .decrypt(new DecryptRequest()
-                                    .withEncryptionContext(testEncryptionContext)
-                                    .withKeyId(testKeyId)
-                                    .withCiphertextBlob(ByteBuffer.wrap(testEncryptedKey.decode()))))
-                            .thenThrow(temporaryKMSException);
-
-                    return temporaryKMSException;
-                }
-
-                @ParameterizedTest
-                @DisplayName("should throw RemoteKeySourceException.")
-                @ValueSource(classes = {
-                        DependencyTimeoutException.class, InvalidGrantTokenException.class,
-                        KMSInternalException.class})
-                void shouldThrowRemoteKeySourceException(final Class<AWSKMSException> invalidKeyExceptionClass) {
-                    final var invalidKeyException = parameterizedBeforeEach(invalidKeyExceptionClass);
-                    assertThatThrownBy(
-                            () -> kmsSymmetricDecrypter.decrypt(testJweHeader, testEncryptedKey, testIv, testCipherText,
-                                    testAuthTag))
-                            .isInstanceOf(TemporaryJOSEException.class)
-                            .hasMessage("A temporary error was thrown from KMS.")
-                            .hasCause(invalidKeyException);
-                }
-            }
-
-            @Nested
             @DisplayName("with a decryption result from KMS,")
-            class WithDecyptionResultFromKMS {
+            class WithDecryptionResultFromKMS {
 
                 private final DecryptResult testDecryptResult = random.nextObject(DecryptResult.class);
                 private final MockedStatic<ContentCryptoProvider> mockContentCryptoProvider =
@@ -269,10 +189,10 @@ class KmsSymmetricDecrypterTest {
                     when(mockAwsKms
                             .decrypt(new DecryptRequest()
                                     .withEncryptionContext(testEncryptionContext)
+                                    .withEncryptionAlgorithm(testJweHeader.getAlgorithm().getName())
                                     .withKeyId(testKeyId)
                                     .withCiphertextBlob(ByteBuffer.wrap(testEncryptedKey.decode()))))
                             .thenReturn(testDecryptResult);
-
                     random.nextBytes(expectedData);
                     mockContentCryptoProvider.when(
                                     () -> ContentCryptoProvider.decrypt(
